feat(status): actionbar with conversion status
Added actionbar, fixed duplicated values in conversion, blocking connectings during a conversion, synchronized TimedPlayer, fix TopOnlineTimeInventory back inventory item, fixed TimedPlayerAdapter

+++ src/main/java/com/nextplugins/onlinetime/dao/adapter/TimedPlayerAdapter.java
@@ -10,7 +10,7 @@ public final class TimedPlayerAdapter implements SQLResultAdapter<TimedPlayer> {
     public TimedPlayer adaptResult(SimpleResultSet resultSet) {
 
         String userName = resultSet.get("name");
-        long timeInServer = resultSet.get("time");
+        int timeInServer = resultSet.get("time");
 
         TimedPlayer timedPlayer = TimedPlayer.builder()
                 .name(userName)
